Add LaTeX template
Add .tex files, which have to be compiled. This is to be used with TeX Studio or another IDE that compiles the thesis.tex file which includes all other tex Files

This is still a WiP and far from finished.
Also added gitignore to contain all the temp files LaTeX creates